wormhole: initial container attach impl
transcript of why I chose to do this:

The problem with docker run, so there are 3 options, docker exec, docker run, and custom linux magic. The problem with docker run is that you can attach the most namespaces, but in order to copy the security properties, and specifically seccomp, you need to start the process in the container with those security properties. And that seccomp filter, assuming we don't have capsys-admin, that seccomp filter blocks unshare, which makes it impossible for me to do the mountain namespace magic. I could start it and choose the job capabilities, but that still makes the seccomp filter too permissive, which is better than nothing, but still a security hole.

Another problem with docker run is that it could result in dangling namespaces. So if you start the original container, the external debug shell, and somehow the original container exits, but the debug shell is still running, now you have a problem. But on the flip side, one advantage of this method, I would need to test, but it might be possible to use this method to debug a stopped container, whereas for my method it's impossible. Because I would not, in that case I would have to use the docker run trick, but only in the stopped container case. However, I think stopped container support is not that important right now.

The second option is docker exec. This is perfect, because it copies all of the security properties perfectly, as well as cgroups and all other isolation. But the problem is that with docker exec, it's impossible for me to smuggle file descriptors in. Maybe I can use ptrace to inject FDs. I can use pidfd to take FDs, but I can't smuggle new FDs in unless I use setcop notifier and install FD, which is impossible to attach to an arbitrary process that I didn't spawn. Without being able to smuggle FDs in, there is no way I can call moveMount to mount these FDs. I also can't call unshare because it's subject to the same problem with setcop, which leaves only the custom method.

Extremely annoying, and a little bit flaky because you have to dump the filter, but I also rely on the assumption that there's only one setcop filter attached, and that machines have no setcop filter, which is true now. But even if it weren't true in the future, it's okay, because I would end up attaching a duplicate filter, which is not ideal, but not a dealbreaker. It's a little bit hard to implement, but this is the best solution.

Now in order to self-use dpfmount, I have to pass the FDs. The binary is on the host for security reasons, and docker machines are not allowed to do this, which means I have to make an sconguest rbc call, and then use FDX to pass the FDs. So let's do this in two calls. The first call to the agent, I get back FDs of the Vue FS, allowing me to mount the Fuse dpf. Then I make a second call to the agent, passing the FD of the Fuse mount, which it can then call moveMountR, and it will pass this FD to the spawn child that is supposed to do the FD magic. As a safety measure, it's also possible to use close-range cloexec, but that's not necessary. So there are a total of four FDs that need to be propagated. Standard in, standard out, and FuseMount.