virtue: hvf: don't inherit VM memory mappings on fork
large mappings (e.g. 64G) cause fork to take ~1 sec

we tried to use posix_spawn to fix this in vmgr by avoiding fork, but
there's still one remaining usage of exec.Command in vmgr due to lack of
setctty support. this is an easier, guaranteed fix that we should keep
even after a full migration to posix_spawn.