virtue: vmm: don't boot until all vCPUs are created
all vCPU threads are left free to race once spawned. we need to wait for
secondary vCPUs to be created, to register with GIC/APIC (happens when
HVF vCPU is created on x86), and to register with the parker before it's
safe to start running guest code. otherwise there could be a race causing
SMP bringup to fail

to fix this, give every vCPU a boot sender/receiver

this might not what actually caused
https://github.com/orbstack/orbstack/issues/1180, but it should still be
fixed: