virtue: x86: gracefully fail on native MSRs to support older gen4/5 CPUs
crashes on 4th and 5th gen with VcpuSetRegister. both write_reg and enable_native_msr return that error (fixed) but the write_reg calls are all standard x86, and so are most of the MSRs.

only SPEC_CTRL, PRED_CMD, XSS are new. include TSX_AUX too to be safe