vinit, rootfs, vmgr: create empty 1G swap.img at runtime
it's simpler to just create this at runtime.

partitions aren't necessary: we never used zram writeback, so remove it.

also, while we're at it, reduce swap to 1G since it's only supposed to
be for OOM emergencies. there's no reason for us to use swap instead of
relying on macOS memory compression and swap in real memory pressure
situations.

- more clear errors for cases like https://github.com/orbstack/orbstack/issues/1241
- apparently bsdtar can race? got "unable to find swap header" on one
  boot