virtue: Ensure that `EventFd`s have the same semantics with edge-triggered readiness polls
`mio`—unlike `libkrun`'s `Epoll` abstraction—is edge-triggered by default. This means that, if the signal is asserted more than once at a time but the subscriber only handles it once, the subscriber may never be woken up again since it will never see the input interest be reasserted.

We fix this by always having `read` consume the buffer until doing so would cause it to block.