virtue: virtiofs: readdirplus: ignore EBADF on parent nodeid
https://github.com/orbstack/orbstack/issues/1275

there are very few nodeids call sites that can actually return EBADF.
refresh_nodeid doesn't count because its error is ignored.

this is the only new one, and it could break. consider:

- create a dir
- parent_nodeid is set to its parent dir's nodeid
- macOS renames it (moves it) into a different parent
- virtiofs server's parent_nodeid is stale
- either side deletes the old parent
- linux FORGETs the old parent
- readdirplus tries to get parent_nodeid = EBADF

in reality, this is *really* hard to reproduce because linux doesn't
FORGET the old parent dentry for a long time and drop_caches doesn't
quite do the job, but it could happen on large projects

parent_ino is not important, so just ignore it.