scon: redirect port 80 for all source IPs
we used to only redirect port 80 if connecting from macOS, but it's hard to propagate the "source OS" info to containers: marks don't persist, and we need to masquerade the source IP seen by users' servers so that it's more intuitive. NAT64 would require reserving 3 IPs in each docker subnet (.0 for macOS v4, .1 for Linux machine, .254 for macOS NAT64) if we want unique IPs that fall within the bridge subnet (which people use for filtering). it's really complicated
redirecting from all source IPs is also more intuitive, as people expect cross-container domains to work just like they do from macOS.

the only case in which this could break is:
  - container has both a port 80 and a non-80 server, both HTTP
  - a client container, or a healthcheck, quickly makes requests to the port 80 server while the server container is still starting
  - the non-80 server starts before 80
=> requests go to the wrong server temporarily
but this should be a very rare case, compared to the advantages of redirecting for all source IPs.