scon: fix unmounting wormhole-overlay when a process is still using it
this commit also includes changes to make wormhole-attach work with rebased changes.
the problem was solved by returning an exit code from wormhole-sttach when it cannot delete /nix.
then, we only unmount wormhole-attach if we do not see that exit code in scon.
this commit also includes some minor refactoring