virtue: hvf: balloon: switch to pmap + anti-double-accounting method
(method 1.5 in Linear)

PURGABLE_EMPTY is slow, and it creates way too many chunks: per-page
granularity is 1.5M+ chunks for 12G memory. XProtect uses the same API
as vmmap, and it suspends the process and gets info on all memory regions,
which is O(N^2), freezing the process for over an hour.

also, running `footprint` on this many regions causes a kernel panic. not
the end of the world since it's a debugging tool, but still bad.

unfortunately this isn't sustainable...