vmgr: Implement `ReadyEvents` service
This service allows the host to notify us when a service on the guest has booted. This is useful for HID since HID forwarding is easier to implement as a guest server than as a guest client.

`newtwork` is an intentional misspelling of network to make it clear that it's another Riley-Certified helper library. Sorry!

[danny: cherry-picked from riley/hid]