virtue: balloon: don't park vCPUs
parking is what causes the high tail latency (up to 15ms! and 3ms for
even small numbers of ranges)

now latencies tend to be bounded to 2.5ms for 256 ranges

instead of parking, let's just catch data/insn aborts and put the
respective vCPUs on a barrier, so that we get on-demand parking only if
vCPUs are running.