virtue: hvf: panic on unhandled MMIO data abort
on data aborts (arm64) or EPT violations (x86), we currently assume that
the abort was caused by a MMIO read/write, and send it to the MMIO bus.
this works for MMIO, but if the address isn't within the range of any
MMIO devices, we still assume it succeeded. garbage (whatever was left
in mmio_buf) will be returned on read, and writes will do nothing, which
is very much not correct emulation behavior...

this is usually fine because data aborts caused by bad kernel pointers
go to EL1, as they're in virtual address space, and Linux *shouldn't* be
mapping and accessing non-existent physical addresses -- but it could
happen.

fix it by panicking if the MMIO request is unhandled, meaning that it
wasn't system RAM (which wouldn't have aborted) or MMIO (which would've
been handled).

HVF doesn't seem to support injecting SErrors, so just panic the vCPU if
this happens. (on x86 we could inject a #GP through the VMCS, but we
don't have injection working at all...)

SError injection would be ideal so that Linux can handle it like it does
on real hardware, but this is fine too.

* I found this bug while working on balloon and temporarily unmapping
  all guest RAM: instead of getting a data abort immediately, the kernel
  panics from a null pointer deref caused by a garbage MMIO read