virtue: profiler: synthesize host kernel syscall stack frames
I've realized that we don't actually need kernel stacks. PC and x16 give
enough info to tell when we're in a syscall. the only other reasons a
thread would be in the kernel are other synchronous exceptions
(basically only data and instruction aborts, which we can get from
ktrace's MACH_vmfault event) and preemption (which we can get from thread
flags, but it's not really useful info for optimization).

profiling the kernel itself is usually pointless, and pollutes the stack.
we can't optimize it, and spindump suffices for when we do want to see
what's going on inside the kernel.

this is also much lower overhead.