virtue: profiler: use ktrace to sample page faults, safer memory reads
kernel stacks would be nice, but slow and infeasible.

they're actually kind of unnecessary: we don't need to profile kernel
code in most cases; we just need to know *when* the kernel is taking CPU
time.

cases:
- syscall: PC tells us
  - waiting state inside kernel: thread info tells us
- preemption: thread info tells us
- page fault: ktrace tells us
  * Guest_Data_Abort_Lower_EL_Exc_ARM and User_Data_Abort_Lower_EL_Exc_ARM
    are more accurate, but that requires tracing all synchronous
    exceptions, which is too much data
  * also, it's wrong to count all Guest_Data_Abort_Lower_EL_Exc_ARM time
    as MACH_vmfault, because it could be MMIO. so we have to trace both --
    even more data
  * so we use MACH_vmfault
  * this also has events like Guest_HVC64_Exc_ARM but those are usually
    only 0-1 absolute time units (i.e. 0-40 ns), so not useful
- miscellaneous aborts: should never happen; it'd be an error

so this uses ktrace to get MACH_vmfault start/end times in Mach absolute
time, then pretends that we sampled them and injects a stack frame if
samples land on a fault.