virtue: virtio-blk: mmap: handle SIGBUS and exit with I/O error
we could wrap memcpy from mmap with sigsetjmp+siglongjmp for proper
I/O error reporting, but that's slow and requires C code, as sigsetjmp
can return twice and can't be used safely in Rust.

instead, just exit on I/O error. this is arguably better UX than the
current behavior, where btrfs gets remounted as read-only after a few
retries (which are unlikely to succeed), and the VM largely hangs/breaks

we could also use a TLS flag to wrap memcpy calls instead of saving
ranges, but that's slower and could unintentionally catch SIGBUS on guest
memory as well

forwarding tested by changing handler to SIGSEGV and printing nil URLs
using Go fmt.Println. it recovers as expected and prints "<nil>" instead
of segfaulting

also tested with SIG_DFL: it exits with SIGBUS as expected

Mach exception ports are easier in some ways, but we can't do siglongjmp
tricks with them (so that'd require a rewrite), and we'd have to forward
Mach messages to a saved ux_handler port (is that even possible?). async
signal safety still applies when it's in the same process: locks could be
held, etc.