virtue: virtio: remove interrupt_status
to work around the lack of MSIs, virtio-mmio requires the ISR to ack
IRQs (i.e. traditional shared IRQs) to find out whether there was a
config change, vring event, or both/neither. however, none of our devices
emit config changes, and we don't have spurious interrupts, so this is
just a redundant SeqCst atomic on every signal_used_queue call.

on the Linux side, we've been skipping the ack for a while, since it just
adds useless VM exits.

in the future, we'd ideally migrate to PCI + MSI + GIC-ITS to support
multiple IRQ events with no ambiguity. if we really need config changes
with virtio-mmio, it can be added as part of the transport / intc calls,
not on a per-device basis.