virtue: virtio-console: add custom HVC protocol for fast nonblocking writes
[ORBVM][PERF] virtio-console: add custom HVC protocol to reduce spinning

tty, console, hvc, and virtio-console are a mess:

- tty is full of spinlocks, so no sleeping
- as a result, so is hvc
- a simpler `struct console` is needed for printk. it also doesn't allow
  sleeping
- virtio-console just submits and polls for completion. since this wakes
  up a thread in krun, it means we spin for 30+ us on the vCPU thread for
  every hvc write (which is empirically ~8 bytes)
- if the host pipe is blocked (which could happen since it writes to
  stdout, which redirects to a file), the vCPU will spin for a while,
  both blocking the VM and burning CPU

it becomes really complicated to do a hybrid sync/HVC device on the VMM
side, so let's just add a simple hypercall as a fastpath for when the
pipe has space, and fall back to virtio+poll when it's full. since these
are non-blocking pipe writes on the host side, it's not much slower than
waking up the virtio-console worker thread, so this is significantly
faster and simpler.

for userspace writes, we can actually use vports (non-hvc ports), which
block and wait for an IRQ to wake the waitqueue. for perf and CPU, keep
using the HVC fastpath in the common put_chars function, but allow
userspace writes to fall back to IRQs.

this makes the console situation pretty good:
- printk writes will use HVC fastpath, then spin as a last resort
- userspace writes will use HVC fastpath, then wait for IRQ

(it was/is possible for Linux to disable the virtio-console IRQ for
not being handled too many times, because it races with the polling in
put_chars, and put_chars could reclaim the used buffer before the IRQ
arrives. this is still a potential issue, but there's no way it'll happen
enough times if we mainly use HVC)

to test: add time.Sleep(100 * time.Millisecond) to vmgr/log_collector.go