virtue: vstate: only run HVF vCPU with gruel waker active
run_emulation does MMIO and HVCs, which could take a while. avoid
unnecessary wakeup syscalls when we're not actually running the HVF vCPU.

note: it'd be a bit more optimal to move this into HvfVcpu so that it
only wraps the hv_vcpu_run call, as sysreg reads can take a
non-negligible amount of time in HvfVcpu::run. however, that starts
getting complicated as there's technically a correctness issue with
a stray pending unterrupt that should be cleared before the next
hv_vcpu_run.