virtue: hvf: balloon: fix double accounting after REUSE
from the balloon doc:

"
This also applies to madvise(REUSE) -- it doesn't do anything if pages
aren't in host pmap. REUSE is even more annoying, because if you prefault
and then REUSE, all pages are now accounted to phys_footprint thanks to
the host pmap! And double accounted, because they'll also be accounted to
the VM pmap on fast fault. So then you have to pmap_remove them, which
can, again, only be done reliably by remapping (due to swap messing with
the protect(NONE) case). The other problem with NONE is that it could
cause SIGSEGV if racing with an access, though we can now handle that
using signal handlers on the host.
"