vmgr: rsvm: fix tracing_subscriber panic if pty returns EAGAIN
if the pty buffer is saturated, tracing_subscriber tries to print the
EAGAIN error (from writing to stdout) to stderr using std::eprint. but
it's the same pty so it can return EAGAIN again, and libstd's stdio
panics on any error.

not an issue in release because vmgr's stdout/stderr are usually a log
file, and files are always blocking

so just make the pty blocking to fix this.