virtue: hvf: fix potential vtimer deadlock in PvlockPark
since pvlock park waits on all signals, we do already handle device IRQs
and IPIs correctly: even though the guest disables IRQs before the
hypercall, it'll re-enable them immediately upon re-entry and handle
them if IRQs weren't disabled before the pvlock wait.

however, we also need to check CNTV_CVAL_EL0 and set a deadline so that
the vtimer IRQ fires on time. otherwise, if the spinlock holder is on
the same vCPU but was preempted, and we try to acquire the same spinlock
on the same vCPU, it'll deadlock forever (or until an IPI or IRQ happens
to interrupt it, which may be never). this situation is normally bad
for performance but will not result in a deadlock because the vtimer IRQ
will fire and trigger scheduler preemption, and the lock owner will get a
chance to run.

in theory, this should be impossible on our kernels because we don't
build with full kernel preemption support, but there's probably an edge
case that makes it possible.