vmgr: netpose: add interposer to fix invalid socket write return values
When pf blocks a packet in ip(6)_output_list, it returns -EJUSTRETURN
in an attempt to return success to the caller. However, ip(6)_output_list
just propagates the returned error from pf_af_hook, which is -EJUSTRETURN.
This gets propagated all the way to the userspace syscall return path, where
-EJUSTRETURN means to not modify userspace registers on return.

Since a successful syscall's return value is in x0 (arm64) or RAX (x86_64),
callers think that the syscall succeeded with a number of bytes written equal
to the last value in x0/RAX. On arm64, this is the fd number, as arg1 is in x0.
On x86_64, this is the syscall number (x2000000 for BSD syscalls).

If bytes written > input length, Go panics with "invalid return from write"
in an attempt to provide WriteAll-like semantics for TCP streams in pfd.Write.
Rust's write_all will also panic with a slice index out of bounds error.

To fix this, check if the return value is the previous value of x0/RAX on the
respective architecture, and if so, clamp the max possible number of bytes
written in order to prevent panics.

This behavior is correct because pf_af_hook returning -EJUSTRETURN is
supposed to pretend that the packet was successfully written, even though
it was dropped.

Details in notes/docs/udp-write-panic.md

https://github.com/golang/go/issues/61060

Fixes: https://github.com/orbstack/orbstack/issues/1334
Co-authored-by: Emmie <emmie@orbstack.dev>