vmgr: vnet: propagate TTL for TCP
we currently propagate TTL for UDP but not TCP. TTL is another way to
prevent routing loops between L3 forwarding-based vmnet bridges and
eth0/gvisor, as the TTL will be exceeded after ~64 hops.

unfortunately this doesn't make TCP traceroute work because we'd have to
fix the ICMP identifier and undo NAT in the embedded TCP header.

it's also possible that TTL changes after the SYN, but that's pretty
unlikely.