vinit, vmgr: update disk quota on boot
if the last boot left quota 100% filled, vinit can panic with ENOSPC when
mounting the wormhole overlayfs, because overlayfs deletes the workdir
on boot, and btrfs metadata is always COW.

instead of moving overlayfs mounts to vinit, just solve the problem once
and for all by updating the quota very early, right after mounting data.
this way, it's OK for vinit to do any kind of data initialization.

Fixes: https://github.com/orbstack/orbstack/issues/1130