Revert "rootfs: enable big 16k pcluster to improve erofs compression"
This reverts commit 04db73ea22aea9f05382fa43503ab4d40fce0560.

In 75c171da8 ("rootfs: update base alpine dependencies to match new packages")
we fixed this error appearing only in release builds:

🐧 system | dnsmasq: bad command line options: unsupported option (check that dnsmasq was compiled with DHCP/TFTP/DNSSEC/DBus support)
📦 scon | ERRO[09-02 11:37:07] dnsmasq exited with error                     error="exit status 1"

... but it came back in v1.7.1 release builds.

it turns out that library/package version mismatches were *NOT* the real
cause of the problem. it's an erofs bug: either creating bad images with
big pcluster enabled, or the kernel driver not reading it correctly.

not worth debugging -- just revert it.