wormhole: track wormhole session lifetime by using monitor and subreaper
previously, we had no way to know if a wormhole session had ended;
background processes could be alive after everything wormhole-related
died. this made it impossible to know when it was safe to unmount
wormhole for nuking. now, we keep the monitor and subreaper processes
alive as long as any process runs in the wormhole session.

we consider it an invalid state for processes from the wormhole session
to persist after the monitor has died. to enforce this, the monitor kills
every process holding on to the wormhole mountns before exiting.

because the subreaper and monitor stay alive after the payload dies, we
now send the exit code of the payload through a pipe. we also set
subreaper and monitor stdout and stderr to a pipe so they don't crash
when the ssh session's stdout and stderr die.

since the monitor has no way of knowing what the payload's pid is in its
ns, we open a socket between the monitor and subreaper that allows the
monitor to communicate when a signal needs to be forwarded.