virtue: gicv3: support priority != 0xa0 to fix Linux 6.11 crash
Linux 6.11's GICv3 driver sets priority to 0xc0 instead of 0xa0.

we don't care what the actual priority value is; we just need to make
sure that all interrupts are configured with the same priority, because
we don't implement prioritization.

enforce that by recording the first requested priority and asserting that
future ones match.