Revert "Revert "vmgr: don't forward packets directed to the vnet subnet""
This reverts commit 68fe50a59d3839f1dc7a51df641e8297a17c6778.

Apparently this commit is fine, and it was an updater issue???

I used Sparkle's BinaryDelta CLI tool to apply all v1.7.4 deltas and
diff patched results against the v1.7.4 release build: everything matches,
so I have no idea what's wrong.

could be a bug only in Sparkle.framework?

whatever it is, this commit is clearly *not* the actual cause, so add it
back.

(I also accidentally shipped `policy drop` for the forward chain too
early, but again, updater means that it shouldn't be the cause.)

https://github.com/orbstack/orbstack/issues/1471