vmgr: routemon: ignore OpenVPN 0./1 and 128./1 default routes
OpenVPN does this as a hack to get more specific default routes on Linux:

Destination        Gateway            Flags               Netif Expire
0/1                10.0.11.1          UGScg               utun4
128.0/1            10.0.11.1          UGSc                utun4

https://serverfault.com/questions/312860/why-openvpn-use-network-0-0-0-0-netmask-128-0-0-0-as-a-default-route

it's not relevant on macOS but is still the default behavior.

since the 128.0.0.0/1 route isn't RTF_GLOBAL, we detect it as a
conflicting "subnet", inhibiting bridge creation.

to fix this, treat /0 and /1 netmasks as RTF_GLOBAL.

also: replace our custom prefixIntersects with (netip.Prefix).Overlaps as
@os-emmie pointed out

Fixes: https://github.com/orbstack/orbstack/issues/1499