scon: bind mount sealed memfd to fix runc 1.1.15 perf regression
runc 1.1.15 regresses container start time because it uses a more
heavyweight CVE-2019-5736 workaround: instead of using a read-only bind
mount of /proc/self/exe, it makes a sealed memfd copy of the runc binary.
the ro bind mount worked fine for us, but they changed it because systemd
enumerates mountinfo on every (un)mount in order to create mount units
and adjust propagation, and that was causing namespace_sem contention and
high CPU usage. doesn't apply to us, but runc still has the fix.

this brings container start time down from 245 to 210 ms, which is
also slightly better than runc 1.1.14 (but barely since we weren't
affected).

ref: runc's contrib/memfd-bind

(for future reference: CVE-2019-5736 is because a container's entrypoint
can be /proc/self/exe, causing runc to execute itself inside the container
despite normally not entering the pid ns. since PR_SET_DUMPABLE is reset
on exec, /proc/self/exe becomes openable in the container, as RW if uid=0.)