swift, vmgr: prevent launch on macOS 15 beta 1–4
macOS 15.0 beta 1–4 has a mismatching ABI for actor-isolated Swift
Concurrency APIs. The new isolated APIs are used when compiling with
Xcode 16+ (swiftc 6.0+).

Betas are considered 15.0 for @backDeployed version checks, causing the
mismatching ABI to be used. This causes segfaults in runProcess,
JsonRPCClient (via AsyncHTTPClient), and more.

We tried working around this in 619a5aa82 by replacing runProcess' use of
withCheckedThrowingContinuation with withUnsafeThrowingContinuation
(which didn't get an actor isolation ABI update), but it turns out that
withTaskCancellationHandler and other Concurrency APIs are also broken,
so it's not feasible to keep working around this by patching libraries.

vmgr also crashes in some unknown Swift code.

add a version gate instead: prevent vmgr startup, and show an NSAlert
on app launch.