scon: kill wormhole session when a stopped container starts
in prepration for using the real overlayfs in stopped container wormhole
sessions (for live-updating reads and writes), we need to enforce
mutual upperdir/workdir exclusion to prevent overlayfs UB.

to do so, use fanotify permission events (FAN_CLASS_PRE_CONTENT) to
intercept dockerd's open() of workdir (FAN_OPEN_PERM). these are
synchronous events, allowing us to block dockerd while we take the time
to kill the temp container.

using the docker API to kill and wait for the temp container could cause
deadlocks if there are global storage driver locks or other
low-granularity locks in dockerd, because we're blocking a dockerd
container start and storage mount. so kill the container's pid 1 instead,
and wait for it to be reapable. (we can't actually reap it because it's
not our child.) the kernel kills and reaps all of pid 1's children before
finishing pid 1's exit process.

https://github.com/torvalds/linux/blob/d73d00352145fb51d31771047aa939850d87fa50/kernel/exit.c#L612

this refactors the wormhole startup and monitoring code path, so that we
can be more careful with fd lifetimes and close rootfsFd (among others)
as soon as possible, to prevent UB warnings from unnecessary mount fds.