scon: domainproxy: propagate RST using nfqueue to fix happy eyeballs
happy eyeballs doesn't work with the TLS proxy because we accept all
connections unconditionally (no Winsock WSAAccept with predicate :c) and
return 502 Bad Gateway on failed upstream connections, but happy eyeballs
needs either a timeout or an RST in order to fall back to IPv4.

to simulate conditional accept4(2), use nfqueue to punt packets to
userspace for firewall decisions. userspace will either accept the packet
(if there's an upstream 80/443 server) or reject it (if upstream RSTs),
propagating the RST to the client.

tproxy BPF can't do this because sk_assign isn't sleepable.

Co-authored-by: Emmie <emmie@orbstack.dev>