scon: fix incorrect partial checksums for tcp and udp, remove nfqueue gso
before, incorrect partial checksums resulted in us having to set the gso
flag for nfqueue so we could avoid the checksum being modified on our
packets. however, this hack didn't work when the partial checksum was
actually correct, so we needed two different queues depending on whether
a packet was coming from mac or from a machine. this is obviously not
ideal.

supposedly, our incorrect partial checksums also caused issues elsewhere.
note that we no longer set the data valid flag on packets from mac.

as it turns out, the partial checksum linux expects is just the negative
of the checksum of the ip pseudo-header. this way, when it takes a
checksum over the entire packet, the pseudo header gets factored in.