vmgr: symlink ~/.local/bin cmds to ~/.orbstack/bin cmds if they exist
otherwise, any links in ~/.local/bin don't update, which is not ideal.
we don't want to just delete them -- that makes it impossible for users
to make those symlinks if they want them. this behavior should be
relatively reasonable. we check if there's .app/ in the old link dest so
we don't overwrite user symlinks to other things. we could check for
OrbStack.app in the old link dest but that would break if some weird user
renames their orbstack app.