scon: use real docker socket in agent
waiting for the proxied docker socket instead of the real one was
causing issues. this fixes that, and should provide a performance
improvement for many agent actions. however, stopped wormhole will not
add certificates now because it is not using the proxied socket. in the
future we should either use the proxied socket or make docker client call
the add certificate function itself.