scon: cert injector: rewrite using runc wrapper for reliability
after all the refactoring, cert injector still has issues:

🌸 agent:docker | time="12-08 01:48:42" level=error msg="failed to unmount merged dir after adding certs" error="device or resource busy" mergedDir=/var/lib/docker/overlay2/2c4faa2c894218cbdfa68c0d381f72aed8c3d3858d061be9f10ed0723da39a6e/.orbstack-merged
🌸 agent:docker | time="12-08 01:48:42" level=error msg="failed to remove merged dir after adding certs" error="remove /var/lib/docker/overlay2/2c4faa2c894218cbdfa68c0d381f72aed8c3d3858d061be9f10ed0723da39a6e/.orbstack-merged: device or resource busy" mergedDir=/var/lib/docker/overlay2/2c4faa2c894218cbdfa68c0d381f72aed8c3d3858d061be9f10ed0723da39a6e/.orbstack-merged

on second thought, the runc wrapper is actually a pretty nice hook point.
it takes care of mounting (regardless of whether containerd snapshotter
is is enabled) without UB (because dockerd is doing it), synchronization
(there won't be more than one "runc start" attempt at a time), checking
for running state (container must be in a valid state if it's being
started), works regardless of start reason (even if it's restart on
engine start), and more.

we decided to use the proxy in order to get labels, but it turns out that
that's pretty easy to do by reading config.v2.json from dockerd. we also
needed mounting info, but mounting is already done by the time "runc start"
runs, and we can just read the rootfs path from the OCI config. (this is
needed because the overlay2 layer ID is different from the container ID)

runc wrapper is also pretty clean: it's a natural extension point of
dockerd/containerd (the OCI runtime), not a hack bolted onto other parts
of the engine.

advantages:
- code is much shorter and simpler
- faster
- supports containerd snapshotter
- no overlayfs UB with this repro:
  while true; do for i in {1..10}; do docker start parasite & docker kill parasite & done; sleep 0.5; done
- works for existing containers created before cert injector was added.
  old impl relies on proxy /start so it misses old
  restart=always/unless-stopped containers
- doesn't leave .orbstack-merged dir behind on crash
- doesn't do anything unusual with overlayfs mounts, so no EBUSY on unmount
- doesn't break docker container deletion (due to EBUSY on rm -fr) if
  we get EBUSY on unmount

a few other changes while I'm at it:
- to remove containers from cache in PreStart, use slices.DeleteFunc
  instead of copying the containers slice, for efficiency
- read files and check contents before writing. faster than writing
  unconditionally; writes always cause a change on disk, have visible
  side effects due to mtime update, and are more risky if UB occurs
- to reduce syscalls, just attempt to write files into directories, and
  rely on open(O_CREAT) returning ENOENT if the dir doesn't exist, or
  ENOTDIR if the parent is a file instead of a dir.