wormhole: exit monitor on ECHILD when reaping
another bug where the wormhole-stub unmount races execing the entrypoint
uncovered this bug, where the monitor hangs if the subreaper gets reaped
before it can reap it. while this shouldn't really happen in normal
cases, it's good to fix it. if reaping gets us ECHILD we know the
subreaper is dead anyway, so it makes sense to cleanup.