wormhole: stub: fix race on start, exit if client crashes early
sometimes stopped wormhole crashes on start because wormhole-attach
unmounts the stub after runc forks and becomes pid 1 in the container,
but before it execs the stub, resulting in:

exec /dev/shm/.orb-wormhole-stub: permission denied

to fix this, make wormhole-attach wait for the stub to start.

also take the chance to fix the issue where we leak an ephemeral
container if the client crashes before attaching: make stub wait for
a connection from attach for up to 5 minutes, and then exit as soon as
monitor does. (pidfd_open is bad because it allows user code in the same
pid ns to SIGKILL monitor, which causes UB)

use the same inotify + stat trick to wait for the stub socket to start
without racing.