scon: dmigrate: fix security issue with docker remote context
for the duration of migration, anything running in the docker machine can
access the remote (ddesktop) docker.sock. this is "fine" because the
forward gets disabled after migration, but if the user kills the
migration with ^C, the forward will be left on indefinitely until vmgr
restarts.

fix this by removing docker machine access entirely, and dialing in scon
instead.