scon: allow canceling clones via mutation hold manager
c.mu.RLock() is safe but causes most other operations to hang for a
very long time while a machine is being cloned.

instead, release the lock and add "mutation holds" to prevent the source
machine from being started or renamed ("mutations") while a clone job is
busy reading from its rootfs ("hold").

also hook clone up to the job manager so that it can be canceled when
either the source or dest machine is deleted.