rootfs: default to btrfs simple quotas on new installs
no reason not to use this on new installs: it's faster (old qgroup kept
showing up in kernel profiles; this doesn't) and does everything we
need it to.

the difference is that old qgroups are hierarchical and require walking
the qgroup tree on every block commit -> high overhead and potential
inconsistency. in squota each extent is owned by exactly one qgroup, and
hierarchical qgroup usage is just calculated as the sum of its descendents,
making adjustments instant (no rescan needed). in exchange, we lose the
ability to correctly track shared vs. owned extents in cases where an
extent first belongs to subvolume A, which is then snapshotted into B,
and then A is deleted. but this shouldn't be a significant issue for us,
even with future machine cloning and snapshotting support.

old installs are harder to deal with because while qgroups are easy
to disable, squota doesn't support rescanning, so all old disk usage is
unaccounted. we might have to implement rescan