rootfs: use new qgroup hierarchy on new data images
I used to not understand how btrfs quotas worked, so all old data images
have shipped with legacy qgroups enabled, and the one default root
subvolume qgroup (0/5). The dynamic disk limit was applied to qgroup 0/5.

This comes with a few issues:
- we did not use subvolumes because they bypass the "global" quota which
  is, in fact, only applied to the root subvolume
- user-created subvolumes bypass our quota calculations entirely

to allow use of subvolumes, initialize new data images with a squota-based
qgroup setup where 1/1 is a top-level group, and all subvolume leaf groups
(0/5, 0/*) are children of 1/1. then vinit can set the limit on 1/1 instead
to make it truly global.