vinit: remove limit from legacy 0/5 qgroup if 1/1 exists
if a user with a new data image downgrades to a pre-1/1 version, vinit
will add a limit to the 0/5 qgroup. after re-upgrading, we need to clear
that limit so that only 1/1 is responsible for limiting.

this doesn't fix the 1/1 limit sticking around on old images, but hopefully
that's not a huge problem: usually more space is used over time, not less,
so the 1/1 limit should be greater than the up-to-date 0/5 one, and users
shouldn't be downgrading for long periods of time.