scon: export: include list of btrfs subvolumes
machines aren't *supposed* to have subvolumes, but this is required to
keep dockerd working if it was using the btrfs storage driver

doesn't preserve snapshots, but it's good enough

it's also not worth the effort to preserve the read-only flag:
we'd have to make a temporary subvolume, then snapshot it. if we create
it as read-only, then we can't extract its contents.