pathfs: build path onto thread-local Vec
This patch makes node path building a lot faster by using a thread
local Vec.

It also includes a rework of the path API, removing the `path` and
`subpath` methods and introducing `with_path` and `with_subpath`, which
remove the need to return cloned `CString`s to callers, and we can
instead let callers temporarily borrow a reference to the used buffers.

Combined with the fact that we're building the path onto thread-local
buffers, this means that a caller can't borrow more than one path at a
time. To address cases where this is necessary, such as renames, etc.,
we use two buffers, allowing up to two concurrent
`with_path`/`with_subpath` calls. More than two concurrent uses of these
result in a panic.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>